refactor(lints): Move missing_features diagnostic to its own rules mod
